third_party/WebKit/LayoutTests/fast/regions/region-styling/webkit-region-error-recovery-expected.txt


PASS One invalid and one valid rule 
PASS One invalid @-rule of allowed type and one valid rule 
FAIL One invalid @-rule of disallowed type and one valid rule assert_equals: Rule valid and in CSSOM. expected 1 but got 0
